refactor(move): helper functions to util.go; consts to constant.go
