+++ constant.go
@@ -1,0 +1,35 @@
+package gorequest
+
+// HTTP methods we support
+const (
+	POST    = "POST"
+	GET     = "GET"
+	HEAD    = "HEAD"
+	PUT     = "PUT"
+	DELETE  = "DELETE"
+	PATCH   = "PATCH"
+	OPTIONS = "OPTIONS"
+)
+
+// Types we support.
+const (
+	TypeJSON       = "json"
+	TypeXML        = "xml"
+	TypeUrlencoded = "urlencoded"
+	TypeForm       = "form"
+	TypeFormData   = "form-data"
+	TypeHTML       = "html"
+	TypeText       = "text"
+	TypeMultipart  = "multipart"
+)
+
+var Types = map[string]string{
+	TypeJSON:       "application/json",
+	TypeXML:        "application/xml",
+	TypeForm:       "application/x-www-form-urlencoded",
+	TypeFormData:   "application/x-www-form-urlencoded",
+	TypeUrlencoded: "application/x-www-form-urlencoded",
+	TypeHTML:       "text/html",
+	TypeText:       "text/plain",
+	TypeMultipart:  "multipart/form-data",
+}
